requirements.txt
torch
transformers
Pillow
pandas
numpy
scikit-learn
nltk
